Correct method signatures for integer, long

+++ Sources/SmithyHTTPAPI/Bindings/HTTPLabelSerializer.swift
@@ -63,12 +63,12 @@ public final class HTTPLabelSerializer: ShapeSerializer {
         matchNonGreedy(label: label, value: "\(value)")
     }
 
-    public func writeInteger(_ schema: Schema, _ value: Int) throws {
+    public func writeInteger(_ schema: Schema, _ value: Int32) throws {
         guard let label = checkForTraitAndGetLabel(schema: schema) else { return }
         matchNonGreedy(label: label, value: "\(value)")
     }
 
-    public func writeLong(_ schema: Schema, _ value: Int) throws {
+    public func writeLong(_ schema: Schema, _ value: Int64) throws {
         guard let label = checkForTraitAndGetLabel(schema: schema) else { return }
         matchNonGreedy(label: label, value: "\(value)")
     }
